refactored changeState code, remove unnecessary logging
