--- src/scheduler/workers/unpublish/findListsInAnnualReviewForReminders.ts
@@ -1,26 +1,0 @@
-import { startOfToday } from "date-fns";
-import { prisma } from "server/models/db/prisma-client";
-
-export async function findListsInAnnualReviewForReminders() {
-  const today = startOfToday().toISOString();
-  return await prisma.list.findMany({
-    where: {
-      nextAnnualReviewStartDate: {
-        lte: today,
-        not: null,
-      },
-      items: {
-        some: {
-          isAnnualReview: true,
-        },
-      },
-    },
-    include: {
-      country: {
-        select: {
-          name: true,
-        },
-      },
-    },
-  });
-}
